reference/LocalMind-master/pyproject.toml
[project]
name = "localmind"
version = "1.0.0"
description = "100% offline local RAG application with fine-tuning support"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    # FastAPI and web
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "sse-starlette>=2.0.0",
    
    # Database
    "sqlmodel>=0.0.14",
    "aiosqlite>=0.19.0",
    
    # Vector store
    "chromadb>=0.4.22",
    
    # LLM and embeddings
    "langchain>=0.1.0",
    "langchain-community>=0.0.13",
    "ollama>=0.1.6",
    "openai>=1.10.0",
    "sentence-transformers>=2.3.1",
    
    # Document parsers
    "pymupdf>=1.23.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.2",
    "pandas>=2.2.0",
    "python-pptx>=0.6.23",
    "pillow>=10.2.0",
    "pytesseract>=0.3.10",
    
    # Configuration
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    
    # Utilities
    "tiktoken>=0.5.2",
    "aiofiles>=23.2.1",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
finetune = [
    "transformers>=4.37.0",
    "peft>=0.8.0",
    "datasets>=2.16.0",
    "accelerate>=0.26.0",
    "bitsandbytes>=0.42.0",
    "trl>=0.7.10",
]
finetune-unsloth = [
    "unsloth>=2024.1",
]
finetune-airllm = [
    "airllm>=2.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
